AstTest renamed to AstEvaluationTest
