+++ src/test/java/com/mmm/his/cer/utility/farser/ast/OperandAndEvaluationOrderTest.java
@@ -3,7 +3,7 @@ package com.mmm.his.cer.utility.farser.ast;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.contains;
 
-import com.mmm.his.cer.utility.farser.ast.AstTest.StringOperandSupplier;
+import com.mmm.his.cer.utility.farser.ast.AstEvaluationTest.StringOperandSupplier;
 import com.mmm.his.cer.utility.farser.ast.parser.DescentParser;
 import com.mmm.his.cer.utility.farser.ast.setup.MaskedContext;
 import com.mmm.his.cer.utility.farser.ast.setup.TestContext;
